:bug: fix toot chaining bug for custom thread entries
It turns out that it never really worked! Mastodon sends its ids as strings, but in the test we always used ints for all ids.
The database holds the social profile / user id as a _number_ and the thread tail id as a _string_, so some functionality remained.

This caused the "is the reply from me and thus part of the thread" filter to fail, and create issues for travel logs that also contain non-trwl toots.
For trwl-only threads, we always save the tail of the thread in our db, so those threads worked fine.