:sparkles: Remove webhook from database when webhook client has responded with 410 Gone
Background: Some users have a large number of webhooks configured, often for the same URL, i.e. because they uninstalled and reinstalled an app. The old webhook will usually not be discarded, so when webhook events are fired, dozens of requests are sent out.
The webhook client (FCM, APNS, ...) knows if the app is still installed and can respond to us if it isn't anymore.