🐛 Fix: No intermediate stops shown on some international journeys
Fix a bug where no intermediate stops are shown when trying to check into an international train operating in a different timezone.
This is due to the way stopovers after the start station are determined:
The `departure_planned` and `departure` passed from the front end cannot simply be compared as they might refer to different timezones.