feat(ui): add _Update Station Wikidata_ button
This PR adds another button with the ability to re-refetch and update
the station information with new data from Wikidata.

If there is new data in Wikidata, there is no way of updating this data
in the (normal) user UI.

To achieve this the GET parameter `allowUpdate` (`boolean`) was added to
the `POST /api/v1/experimental/wikidata?allowUpdate=1` endpoint.
This parameter prevents the early exit (because the `wikidata_id`
already exists) and so `WikidataImportService::searchStation` is
invoked, which updates the Station database with fresh data from
Wikidata.

I could locally test this feature successfully.

There are some more questions:

- is this feature (request) useful? (if not, just stop reading and close the PR)
- is a GET parameter the correct choice here? (or would another endpoint
  suit better)
   - is the name `allowUpdate` alright?
- is the `allowUpdate` parameter even required or can it's behaviour be
  the default?
- what is the best way to prevent code duplication for the JS function
  `fetchWikidata` (and where to put it then)?

I'm looking forward for some input.